backend/requirements.txt
# Face Recognition API Requirements
fastapi==0.104.1
uvicorn==0.24.0
torch==2.1.0
torchvision==0.16.0
facenet-pytorch==2.5.3
pillow==10.1.0
numpy==1.24.3
python-multipart==0.0.6

# Optional: For GPU support (CUDA)
# torch==2.1.0+cu121 --index-url https://download.pytorch.org/whl/cu121
# torchvision==0.16.0+cu121 --index-url https://download.pytorch.org/whl/cu121